+++ admin/src/components/desk-header.tsx
@@ -192,6 +192,7 @@ export function DeskHeader({
                 width={252}
                 height={48}
                 className="hidden h-12 w-auto shrink-0 object-contain compact:block"
+                style={{ width: "auto" }}
                 priority
               />
             </Link>